+++ codex-rs/chatgpt/README.md
@@ -1,5 +1,5 @@
 # ChatGPT
 
-Code in this crate pertains to first party ChatGPT APIs and products such as Codex agent.
+This crate pertains to first party ChatGPT APIs and products such as Codex agent.
 
-This crate should be primarily built and maintained by OpenAI employees.
+This crate should be primarily built and maintained by OpenAI employees. Please reach out to a maintainer before making an external contribution.
